fix: address PR review comments for class feature selection
- Add localization for all user-facing strings in class feature components
- Use ClassFeatureSelectionProps type instead of inline props
- Remove duplicate ClassFeatureResult type from getClassFeatures.ts
- Add error handling (.catch) to async $effect in state.svelte.ts
- Remove index signatures from NimbleAncestryItem/NimbleBackgroundItem
- Optimize getClassFeatures to skip non-Item packs
- Move utils.ts functions into state.svelte.ts as unexported helpers
- Remove barrel re-exports from types.ts
- Revert unrelated scrollIntoView change
- Fix chevron icon semantics (use chevron-down as default)
